+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 
 Changes in 0.7.X
 =================
+- Updated queryset.delete so you can use with skip / limit (MongoEngine/mongoengine#107)
 - Updated index creation allows kwargs to be passed through refs (MongoEngine/mongoengine#104)
 - Fixed Q object merge edge case (MongoEngine/mongoengine#109)
 - Fixed reloading on sharded documents (hmarr/mongoengine#569)